requirements.txt
clevercsv=0.6.6
regex=2020.11.13
chardet=4.0.0
google-cloud-bigquery=2.6.1